refactor(models): deduplicate RouteFamily enum definition
