+++ lib/models/api/v6/network/interfaces.dart
@@ -1,4 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:pi_hole_client/models/api/v6/network/common.dart';
 
 part 'interfaces.freezed.dart';
 part 'interfaces.g.dart';
@@ -127,19 +128,4 @@ sealed class InterfaceAddress with _$InterfaceAddress {
 
   factory InterfaceAddress.fromJson(Map<String, dynamic> json) =>
       _$InterfaceAddressFromJson(json);
-}
-
-enum RouteFamily {
-  @JsonValue('inet')
-  inet,
-  @JsonValue('inet6')
-  inet6,
-  @JsonValue('link')
-  link,
-  @JsonValue('mpls')
-  mpls,
-  @JsonValue('bridge')
-  bridge,
-  @JsonValue('unknown')
-  unknown,
 }
